pyproject.toml
[tool.poetry]
name = "ws_berezka"
version = "0.1.0"
description = ""
authors = ["ChubAV <chub.aleksandr.v@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
websockets = "^10.3"
selenium = "^4.4.3"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.1"
pika = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
